New folder (2)/civic kural eg/civickural/heck1/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
